+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.0-beta2](https://github.com/cryptomator/hub/compare/1.5.0-beta1...1.5.0-beta2)
+
+### Added
+
+- Allow admins to archive and unarchive any vault (#283, #430)
+- Disable users to exclude them from license seat count (#427, #428)
+- Display a banner to indicate that legacy devices are still in use, since these will be removed in the next major release (#420)
+
+### Fixed
+
+- Creating user with sole role `create-vault` was unable to login due to missing role `user`
+
 ## [1.5.0-beta1](https://github.com/cryptomator/hub/compare/1.4.6...1.5.0-beta1)
 
 ### Added